# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Darwin Ops MCP Backend API with Tool Registry, Policy Engine, and Approval Flow.\nReturns immediate execution result or 202 + approval ID for high-risk tools.",
        "title": "Darwin Ops MCP API",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "support@darwin-ops-mcp.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/agent-keys": {
            "get": {
                "description": "Lists API key metadata. Plaintext secrets and hashes are never returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-keys"
                ],
                "summary": "List Agent API Keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Issues a bearer token for an AI agent. Requires the master API token when authentication is enabled.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-keys"
                ],
                "summary": "Create Agent API Key",
                "parameters": [
                    {
                        "description": "Agent API key issuance request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agent-keys/{id}/revoke": {
            "post": {
                "description": "Revokes an agent API key by ID. Revoked keys can no longer authenticate.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent-keys"
                ],
                "summary": "Revoke Agent API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}